update climada_measures_viewer in salvador_demo
major restructuring, design change, cleanup and enabling of loading results from both San Salvador and Barisal.
The new measures_impact struct that is loaded can contain a not specified number of entries.
The measures_impact struct requires now the fields .scenario and .entity for each EDS struct, so that it is not
necessary to load a separate entity.